requirements.txt
requests
pysocks
mechanize